cherry picked eb1382156b33b7391dd17c8d4eddba0af8a2ee82 to `movement-camera-testing`
moved for organization.
original description:
`made tweaks to the movement system to allow ground friction. Now the physics material of the ground can be changed to affect the player friction instead of always being controled by air resistance. This is a work in progress and isn't finished yet. Also added a gun camera separate from the main camera that just renders the viewmodels. Will be expanded further to make camera animations possible.`