player friction and collider updates
forgot to mention the new player colliders last commit. there are now 2 different colliders for the player. One for the feet and one for the body. The feet have zero friction be default but will get the maximum friction of the ground, so they just use the friction of the ground. The body material has less friction to prevent it from sticking to walls and climbing ramps that shouldn't be scaled.