Gotta stay strapped kids :smiling_imp:
When someone says they don't like you but smile after it, they are usually lying. This is a poor attempt at asserting dominance. DO NOT LET IT SLIDE! If they think they can get away with this they will use you like a doormat. The proper response would be to hit the  griddy as hard a possible followed by a kick to their groin region. This will make sure to put you back on top, and they will know that you are the a unchallenged alpha male.